[core] Replaced default bulk data encryption cipher with AES-CBC+HMAC
This is used in cases where we reuse the same key for encryption but are
unable to keep a counter around (eg: remote contact info). Even though
we use cryptographically secure random IVs each time, there is the
infinitesimal chance an IV may be reused, which could lead to key
compromise. So to be on the safe side, we switch to using
AES-256-CBC+HMAC-SHA256 (with an encrypt-then-MAC scheme over the IV and
ciphertext). Remote contact encryption will continue with AES-256-GCM
until clients have migrated to the new version, otherwise older clients
will no longer be able to read newly updated contact information.